matrix examples header
